requirements.txt
langchain==0.1.0
langchain-community==0.0.10
requests==2.31.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
huggingface-hub==0.19.4
